board: add Oposite method for letter
