+++ ttgboard/letter.go
@@ -1,0 +1,53 @@
+package ttgboard
+
+// IdxState represents index's state
+type Letter byte
+
+const (
+	LetterNone Letter = iota
+	LetterX
+	LetterO
+)
+
+func newBoardIndex() *Letter {
+	result := LetterNone
+
+	return &result
+}
+
+// SetState sets index state
+func (l *Letter) SetState(state Letter) {
+	*l = state
+}
+
+// String returns Letter's string
+func (l *Letter) String() string {
+	switch *l {
+	case LetterNone:
+		return " "
+	case LetterX:
+		return "X"
+	case LetterO:
+		return "O"
+	}
+
+	// should not be reached
+	return "?"
+}
+
+// IsNone return's true if Letter == LetterNone
+func (l *Letter) IsNone() bool {
+	return *l == LetterNone
+}
+
+// Oposite returns letter, which is an oposite to current
+func (l *Letter) Oposite() Letter {
+	switch *l {
+	case LetterX:
+		return LetterO
+	case LetterO:
+		return LetterX
+	}
+
+	return LetterNone
+}
